server/requirements.txt
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
SQLAlchemy==2.0.43
typing_extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2

# Testing
pytest==8.3.3
pytest-flask==1.3.0
